Removes unused OSSEC email setup script
The `setup_gmail_test.sh` script is a legacy artifact, not used anywhere
in the configuration, and not even mentioned in the project docs. Rather
than update it according to shellcheck's linting standards for shell
scripts, let's just remove it. We can always dig it out of git history
if we need it in the future, but as of today, it's simply cruft.