Skip shellcheck on the OSSEC deb package scripts
We have a dedicated issue for cleaning up the OSSEC Debian packages, in
particular the rat's nest of postinst setup lines that contain many
shellcheck violations. Fixing those issues now, during the introduction
of the shellcheck linting target, would complicate review, and belongs
in a separate PR.

Since the find command got gnarlier, prefixing it with `@` to avoid
printing the pipe chain to stdout.