Uses official shellcheck Docker image for linting
The Dockerfile is built from "scratch" and contains only the shellcheck
binary. Therefore we still have to run `find` outside the container and
pass in the filepaths relative to the `/mnt` directory inside the
container, which is also set as the working directory.

Also excluding SC1091 since the container won't have any of the included
files. Better for cross-platform testing anyway, since those checks
would have passed on Linux developer workstations, but failed on macOS
anyway.

Relevant resources:

  * https://hub.docker.com/r/koalaman/shellcheck/
  * https://github.com/koalaman/shellcheck